Avoid incorrect prior visiblity update
Some containers were mistakenly rejected as no-ops when calculate
targets because the mVisible of changes are incorrectly updated to
false in the updatePriorVisibility method.

For example, open app A and go home, then open app B immediately.
We collect the launcher in the last OPEN transition and the origin
mVisible is true of launcher change, then set visible request false
of launcher. But at this time, the first OPEN transition has just
finished and start committing invisble of launcher.
So when the last transition is ready, updatePriorVisibility method
is called and set mVisible false of launcher change,
so the launcher change will not add to the final targets.

Bug: 345649450

Change-Id: I7ca6080f6df537cc032e21ed8fa38e8dcebaf64e
Signed-off-by: tanxiaoyan <tanxiaoyan@xiaomi.com>